Converts all error codes from 0 to -1

+++ abi.h
@@ -197,7 +197,9 @@ typedef struct {
 bool is_tuple_type(ABI_t t);
 
 // Helper to get the number of parameters in a tuple type
-size_t get_tuple_sz(ABI_t t);
+// @param `t`        - Parameter to inspect
+// @return           - Number of tuple params; -1 if this is not a tuple
+int get_tuple_sz(ABI_t t);
 
 // Ensure we have a valid ABI schema being passed. We check the following:
 // * Is each atomic type a valid ABI type? (e.g. uint32, string)
@@ -206,7 +208,7 @@ size_t get_tuple_sz(ABI_t t);
 // meaning things like `string[3][3]` and `string[]` are allowed, but
 // `string[3][]` are not. This is because the ABI spec is pretty loose about
 // defining these encodings, so we will just be strict and reject combinations.
-// @param `tyes`      - array of types making up the schema
+// @param `types`     - array of types making up the schema
 // @param `numTypes`  - number of types in the schema
 // @return            - true if we can handle every type in this schema
 bool abi_is_valid_schema(const ABI_t * types, size_t numTypes);
@@ -218,12 +220,12 @@ bool abi_is_valid_schema(const ABI_t * types, size_t numTypes);
 // @param `info`      - information about the data to be selected
 // @param `in`        - Buffer containin the input data
 // @param `inSz`      - Size of `in`
-// @return            - Size of array dimension; 0 on error.
-size_t abi_get_array_sz(const ABI_t * types, 
-                        size_t numTypes, 
-                        ABISelector_t info, 
-                        const void * in,
-                        size_t inSz);
+// @return            - Size of array dimension; -1 on error.
+int abi_get_array_sz( const ABI_t * types, 
+                      size_t numTypes, 
+                      ABISelector_t info, 
+                      const void * in,
+                      size_t inSz);
 
 // Get the array size of a type inside of a tuple. Must be a variable size array.
 // @param `types`     - array of ABI type definitions
@@ -232,13 +234,13 @@ size_t abi_get_array_sz(const ABI_t * types,
 // @param `paramInfo` - information about the parameter we want inside the tuple
 // @param `in`        - Buffer containin the input data
 // @param `inSz`      - Size of `in`
-// @return            - Size of array dimension; 0 on error.
-size_t abi_get_tuple_param_array_sz(const ABI_t * types, 
-                                    size_t numTypes, 
-                                    ABISelector_t tupleInfo,
-                                    ABISelector_t paramInfo, 
-                                    const void * in,
-                                    size_t inSz);
+// @return            - Size of array dimension; -1 on error.
+int abi_get_tuple_param_array_sz( const ABI_t * types, 
+                                  size_t numTypes, 
+                                  ABISelector_t tupleInfo,
+                                  ABISelector_t paramInfo, 
+                                  const void * in,
+                                  size_t inSz);
 
 // Decode and return a param's data in `out` given a set of ABI types and an `in` buffer.
 // Note that padding is stripped from elementary types, which are encoded in 32-byte words regardless
@@ -251,14 +253,14 @@ size_t abi_get_tuple_param_array_sz(const ABI_t * types,
 // @param `info`      - information about the data to be selected
 // @param `in`        - Buffer containin the input data
 // @param `inSz`      - Size of `in`
-// @return            - number of bytes written to `out`; 0 on error.
-size_t abi_decode_param(void * out, 
-                        size_t outSz, 
-                        const ABI_t * types, 
-                        size_t numTypes, 
-                        ABISelector_t info, 
-                        const void * in,
-                        size_t inSz);
+// @return            - number of bytes written to `out`; -1 on error.
+int abi_decode_param( void * out, 
+                      size_t outSz, 
+                      const ABI_t * types, 
+                      size_t numTypes, 
+                      ABISelector_t info, 
+                      const void * in,
+                      size_t inSz);
 
 // Perform `abi_decode_param` on a parameter nested in a tuple struct.
 // Tuple data is encoded as if it is its own definition and is offset like dynamic data.
@@ -270,15 +272,15 @@ size_t abi_decode_param(void * out,
 // @param `paramInfo` - information about the param inside the tuple
 // @param `in`        - Buffer containin the input data
 // @param `inSz`      - Size of `in`
-// @return            - number of bytes written to `out`; 0 on error.
-size_t abi_decode_tuple_param(void * out, 
-                              size_t outSz, 
-                              const ABI_t * types, 
-                              size_t numTypes,
-                              ABISelector_t tupleInfo,
-                              ABISelector_t paramInfo, 
-                              const void * in,
-                              size_t inSz);
+// @return            - number of bytes written to `out`; -1 on error.
+int abi_decode_tuple_param( void * out, 
+                            size_t outSz, 
+                            const ABI_t * types, 
+                            size_t numTypes,
+                            ABISelector_t tupleInfo,
+                            ABISelector_t paramInfo, 
+                            const void * in,
+                            size_t inSz);
 
 // Encode a payload given a set of types. 
 // All parameter data should be tightly packed in `in`. Numbers are expected to be little endian buffers.
@@ -290,13 +292,13 @@ size_t abi_decode_tuple_param(void * out,
 // @param `offsets`   - list of size `numTypes` containin offsets for the types' data in `in`
 // @param `in`        - Buffer containin the input data
 // @param `inSz`      - Size of `in`
-// @return            - number of bytes written to `out`; 0 on error.
-size_t abi_encode(void * out, 
-                  size_t outSz, 
-                  const ABI_t * types, 
-                  size_t numTypes, 
-                  size_t * offsets, 
-                  const void * in, 
-                  size_t inSz);
+// @return            - number of bytes written to `out`; -1 on error.
+int abi_encode( void * out, 
+                size_t outSz, 
+                const ABI_t * types, 
+                size_t numTypes, 
+                size_t * offsets, 
+                const void * in, 
+                size_t inSz);
 
 #endif
